+++ CHANGELOG.md
@@ -6,5 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.2] - 2021-03-20
+- dtm now uses determinants module for determinant-based subroutines
+- made arrays storing configurations (iconf1, iconf2) consistent between all programs
+- several minor text edits
+- descriptions added for several subroutines in davidson module
+
+## [0.0.1] - 2021-03-17
+Fixed issue with table of J being unordered.
+
 ## [0.0.0] - 2021-03-17
 Original source import.